refactor(constants): remove barrel file and use explicit imports
