--- app/constants/index.ts
@@ -1,3 +1,0 @@
-export * from './redirects'
-export * from './config'
-export * from './sitemap'
